requirements.txt
mlx
mlx-lm
langchain-huggingface
langchain-chroma
langchain-text-splitters
duckduckgo-search
pyyaml
ipywidgets
tqdmfastapi
uvicorn
